Step 6 - Add custom base abstractions
Provided base abstraction for read-only Repositories via the
ReadOnlyRepository interface in this case we also fix the the type of
the ID property to Long. 

Added Product domain class, sample data as well as an extension of
ReadOnlyRepository to declare a query method to find Product instances
containing a given description.